Rename common folder to components
